+++ src/playground-ide.ts
@@ -150,8 +150,8 @@ export class PlaygroundIde extends LitElement {
    *
    * This URL must host the following files from the playground-elements
    * package:
-   *   1. playground-service-worker-proxy.html
-   *   2. service-worker.js
+   *   1. playground-service-worker.js
+   *   2. playground-service-worker-proxy.html
    *
    * Defaults to the directory containing the script that defines this element
    * on the same origin (typically something like