+++ rollup.config.configurator.js
@@ -36,11 +36,15 @@ export default [
             dest: 'configurator/deploy/',
           },
           {
-            src: 'typescript-worker.js',
+            src: 'playground-typescript-worker.js',
             dest: 'configurator/deploy/',
           },
           {
-            src: 'service-worker.js',
+            src: 'playground-service-worker.js',
+            dest: 'configurator/deploy/',
+          },
+          {
+            src: 'playground-service-worker-proxy.html',
             dest: 'configurator/deploy/',
           },
         ],